+++ tests/testthat/test-estimate_carbon.R
@@ -23,7 +23,7 @@ test_that("estimates match those in raw data", {
       CARBON_AG,
       DRYBIO_AG
     )
-  data_prepped <- prep_data(db) |>
+  data_prepped <- fia_tidy(db) |>
     dplyr::filter(INVYR >= 2000L) |>
     rename(YEAR = INVYR) |>
     prep_carbon() |>