+++ tests/testthat/test-interpolate_data.R
@@ -4,7 +4,7 @@ test_that("variables with NAs get interpolated correctly", {
     "DE",
     dir = system.file("exdata", package = "forestTIME.builder")
   )
-  data <- prep_data(db) |>
+  data <- fia_tidy(db) |>
     dplyr::filter(tree_ID == "10_1_1_104_1_28") |>
     dplyr::select(
       plot_ID,
@@ -36,7 +36,7 @@ test_that("interpolation flags negative numbers as fallen dead", {
     "DE",
     dir = system.file("exdata", package = "forestTIME.builder")
   )
-  data <- prep_data(db) |>
+  data <- fia_tidy(db) |>
     dplyr::filter(tree_ID %in% c("10_1_1_104_3_4", "10_1_1_148_4_2")) |>
     dplyr::select(
       plot_ID,
@@ -75,7 +75,7 @@ test_that("interpolation of CULL is correct", {
     "DE",
     dir = system.file("exdata", package = "forestTIME.builder")
   ) |>
-    prep_data() |>
+    fia_tidy() |>
     dplyr::filter(tree_ID == "10_1_1_128_1_24")
 
   data_interpolated <- data |>